reactor: 修改termina/models目录结构
