+++ apps/terminal/models/task.py
@@ -1,0 +1,25 @@
+from __future__ import unicode_literals
+
+import uuid
+
+from django.db import models
+from django.utils.translation import ugettext_lazy as _
+from .terminal import Terminal
+
+
+class Task(models.Model):
+    NAME_CHOICES = (
+        ("kill_session", "Kill Session"),
+    )
+
+    id = models.UUIDField(default=uuid.uuid4, primary_key=True)
+    name = models.CharField(max_length=128, choices=NAME_CHOICES, verbose_name=_("Name"))
+    args = models.CharField(max_length=1024, verbose_name=_("Args"))
+    terminal = models.ForeignKey(Terminal, null=True, on_delete=models.SET_NULL)
+    is_finished = models.BooleanField(default=False)
+    date_created = models.DateTimeField(auto_now_add=True)
+    date_finished = models.DateTimeField(null=True)
+
+    class Meta:
+        db_table = "terminal_task"
+
